This is just a smallish commit so I log my thoughts.
Just as yesterday I know how to do it in Spiral. And if I wanted to make a type system, rather than change Spiral it would be better to make a language that compiles to it.

I will admit, for a while now I've thought of type systems as the enemy. I appreciated the documentation benefits F# brings me, along with the quick IDE feedback, but I've grown to resent it because the work on Spiral probably shaved a year of my life and turned me into an unwilling language researcher.

But yesterday was the first time I really saw type systems as integral parts of reality. I realized that for the first time, even if I put one at top Spiral and did not even track binding time, there is a definite benefit to them.

But even so, I do not think I need them. I can just imitate them by hand in the parser library.

Now, I said I will drop Spiral if I judge LMS to be better, but that was just a thing I said in the spur of the moment. There is no way I can do that - I am really way too deep right now.

http://web.stanford.edu/class/cs442/

What I will do is some more research on LMS, and maybe try to figure out how to deal with its very sparse documentation, how it compiles to the GPU, why do the DSLs immediately jump to the very highest level instead of introducing the low level stuff first and various other questions I have.

Probably the most important questions are why are the ML libraries made with LMS abandonware and why is LMS so unknown in the Scala community.

I said I would abandon Spiral and then rescinded that just now, but if LMS had a vibrant ecosystem with top notch ML libraries around it, then I'd really never probably have started work on it in the first place.

Yesterday I've been thinking of making a post on the ML sub, but if Scala had what I need, no doubt I would have heard about it already on that sub.

So let me go through the slides from that Stanford class.

I still want to rest a bit. No doubt it will take me til the end of August to finish the language.