I understand now. Spiral is special. I've been thinking long about LMS, but it has a crucial flaw of representing dynamic data with Rep. It would be much better if it was the opposite, if Rep were static and the standard things were dynamic as with regular Scala.
I have a bunch of question I would like to ask Rompf such as:

1) Why are Rep types used to represent dynamic data instead of vice versa?
2) What is the degree of integration between LMS DSLs and regular Scala code?
3) Related to #2, as a litmus test, can I pass a regular Scala function to its GPU DSL and have it compiled as a Cuda kernel?

I could ask these questions, but I already know the answers to them based on the context. In all likelihood LMS has poor integration otherwise it would be much more popular. And if they had such a powerful feature as passing lambdas between host and the GPU backends, you can bet they would have advertized it.

And regarding #1, there is no point in asking this question now. The time to ask was 7+ years ago when the work on LMS started. It is too late to fix this design mistake now.

Because Rep is used for dynamic variables, that means if I want to write Scala code in such a fashion, it will be incredibly onerous compared to Spiral. In fact, even if it was the opposite, I'd very much prefer the dynamic kind of staging that Spiral does.

The real reason why LMS is not popular is in fact because it is unwieldy.

Out of all the languages, Spiral is closest to the True One.

That this should be the case in 2017, makes me quite a bit depressed. I mean that literally.

I do not feel like programming, but I need to get my groove back. Small goals. Forget everything that came before. Spiral is everything that I said it was in the last Simulacrum update, but I will need a while before I can make it usable. For tomorrow, I will finally deal with the singleton tuples/union vars issue. I will extend the core language one last time. And then I will dig up the list…

~1) Dynamise~
2) Arrays
3) Cuda
4) Closures
~5) Type constructors for primitives. It is stupid to keep passing 0 when I mean int64 even if it works.~
6) Standard syntax for pattern matching
7) Standard library functions for arrays since F#'s are so slow. I'll need to do it separately for Cuda arrays anyway.
~8) Testing .NET stuff~
9) Modules
10) Split TyV into TyTag and TyV in order to deal with the singleton union issue.

And go at it one by one. I wanted to take a break by doing HackerRank exercises a few days ago, but it seems that won't be possible until I finish all of the above. Right now just parsing a few ints is an amazing pain in the ass. Forget about everything, I need to make the entire language usable. That won't be a quick task, but I think I can manage it in a month if I focus.