~1) Dynamize~ 2) Arrays and references 3) Cuda 4) Closures ~5) Type constructors for primitives. It is stupid to keep passing 0 when I mean int64 even if it works.~ 6) Standard syntax for pattern matching 7) Standard library functions for arrays since F#'s are so slow. I'll need to do it separately for Cuda arrays anyway. ~8) Testing .NET stuff~ 9) Modules ~10) Split TyV into TyTag and TyV in order to deal with the singleton union issue.~
This issue gave so much grief, but it went away so easily. Though, I am wondering now why the previous thing did not work, but it does not matter.

The fact that I forgot about it is proof enough that it was incomplete.

Well, assuming I did not break anything and I probably did not, that takes care of #10 which I said I would do yesterday.

Let me take care of modules next. It won't be hard.