I want to continue, but I had enough. I did a decent bit of work today. After the initial burst which lasted to about 2:15pm from about the time I woke up at 8:15am, I wanted to do more, but I got really mentally exhausted and spend like two hours recharging my batteries.
I half-slacked by posting on the /r/compsci subreddit.

I did maybe an hour of work in between 2:15 and now, but I guess it was a good hour.

I am done with plugging the arrays into the pevaller and the codegen. Tomorrow I will deal with the parser and then move to testing.

I think I will play Nier and watch anime for the rest of the day.

Really, one good burst per day is really the most I can do. I was dreading doing the rest, but it turns out arrays and modules were not a big deal. I was surprised though to figure out how neater array indexing came out this time.

In actuality, I've dealt with all the major obstacles in the language by end of June. Now I am moving into mop up.