Damn, I did not think .NET arrays would have a hard 32-bit limit.
This is giving me some trouble right now since 64-bit ints are default in Spiral.

Also I made it so array sizes can be either 32-bit and 64-bit, either signed and unsigned ints. the same goes for indexing, but I forgot to test that indexes and sizes have equal types.

The way to fix this would be to do conversion, but I can't be assed to implement that right now.

Let me just restrict array sizes to the int64 and the same for indexes. Then I will just cast the index to int32 at runtime.

That will allow the code to compile.

But the 32-bit limit for arrays is kind of pushing it in this day and age. I really want to have larger arrays so I can future-proof the language.

I definitely do not want to restrict the array indexes to 32-bit on the Cuda side.

Well, let me just deal with this and I will take a break.

Edit: For some reason the above did not go through.

Now that I am doing testing, it seems I need to make sure that for higher dimensional arrays I will need to insert extra boundary checking in order to make sure everything is in order.

Let me do it and maybe I will call it a day after that.

This is really slow and steady work.