I only needed to do a bit of debugging. Done with #6.
~1) Dynamize~
~2) Arrays and references~
3) Cuda
~4) Closures~
~5) Type constructors for primitives. It is stupid to keep passing 0 when I mean int64 even if it works.~
~6) Standard syntax for pattern matching. Also add `when` along with adding matching on literals.~
7) Standard library functions for arrays since F#'s are so slow. I'll need to do it separately for Cuda arrays anyway.
~8) Testing .NET stuff~
~9) Modules~
~10) Split TyV into TyTag and TyV in order to deal with the singleton union issue.~
11) Make a special parser for `type`s.

Yeah, I need to work some more on parsing. And that won't be the end of it. Eventually I will want to put in additional operators. I think I will do it like in Haskell rather than F#.

I also need to make make mset a statement.

Well, today it seems like I will be distracted with things other than programming anyway, so for today and tomorrow my mission will be to work more on parsers. After that I will try and make a primitive parser combinator library.

When I have that, I will finally be able to start doing HackerRank exercises in Spiral. Maybe I'll be able to finally relax a bit programming wise with that.

That kind of rest, I really earned after shedding blood and sweat for so long on the language.

And when that break is over, I will complete the Cuda backend and make the Spiral ML library.

After that I will finally be able to enter the fray and do the things I really wanted to do all along.

For now, let me resume work on the parsers.