inl lit_lift x =     print_static "I am in lit_lift."     lit_lift x inl t1 x = -x inl t2 x = `x inl t3 x = .(x) t1 2.2, t2 true, t3 "asd"
This works now. Great.

~1) Dynamize~
~2) Arrays and references~
3) Cuda
~4) Closures~
~5) Type constructors for primitives. It is stupid to keep passing 0 when I mean int64 even if it works.~
~6) Standard syntax for pattern matching. Also add `when` along with adding matching on literals.~
7) Standard library functions for arrays since F#'s are so slow. I'll need to do it separately for Cuda arrays anyway.
~8) Testing .NET stuff~
~9) Modules~
~10) Split TyV into TyTag and TyV in order to deal with the singleton union issue.~
~11) Make a special parser for the `type` keyword.~
~12) Implement F# style user defined binary infix operators.~
~13) Implement user defined unary prefix operators.~

Done with #13. They prefix operators are more in the Haskell style. Haskell only has a single prefix operator which is negate. Spiral on the other hand has 3.

Maybe I should add an operator so functions can be used in infix form as in Haskell? Meh, it is not important.

Now, since I just downloaded Doom and I finished this, I feel like playing it. No way I feel like doing HackerRank exercises now.

Let me play it for a bit and then I will start work on parser combinators. Though I doubt I will be doing it today. That is another big thing to dive into and I do not want to do it when I have only like 1 good hour left in the day.