I added the when and as patterns. I spent a fair amount of time trying to figure out why the when pattern kept running into parser patterns despite knowing the workaround from the start.
In the end, it seems that negate and operators parsers consider -> to be a part of their pattern and that gives rise to problems.

I could built in workaround so that when does not need parenths, but doing it the same F# is not too bad.

With that out of the way, let me continue work on the Spiral parsers library. With the when pattern, I won't need to put in if_static as a keyword.