"Ok, this is it. This is really, really good. What I am going to do next is bring back the tuple operations so that I can a generic tuple function rather than having tuple2, tuple3 etc. like Fparsec has.
Ah, I jumped the gun a bit in |>>. I really need to have it call return. It is not correct at the moment.
inl (|>>) a f = a >>= fun x s ret -> ret (.Succ, f x)

Yeah, like this. This is it. I nearly put the arguments in the wrong place, but by now I am actually fully comfortable with CPS and can do it in my head.

Remember my first meeting with it, that quicksort function that seemed like absolute black magic?

I've come such a long way."