I fixed a few bugs in the char parser and got the Spiral Parsing to parse.
I feel the stimulation leaving me a bit. I need to take a break to plan my next move after all this debugging work.

...

Done with the break.

What I will do is clear my head of stray thoughts and focus on just one thing - making a parser to parse a single int from a string.

The Parsing module that I have right now is just an untested skeleton. I do not have any way of even running a parser currently.

That is what will be the focus of my next move.