I finally got all the bugs out of the way and now that many parser compiles correctly.
Phew. Now one problem that I have is that the compiler is excessively slow. 1.4s for the parser example is quite a lot. I shudder to think how much it would take it to compile at a larger scale.

What should I do here?

Ordinarily I would add dirty flags, but the CSE optimization blogs that. And it is not just an optimization that I can turn off. I am taking advantage of it to communicate in destructuring.

This is troublesome. I am not sure what I will do, but I guess I will try profiling for starters.

After that I will start playing around with HackerRank exercises. Soon it will be time for my long awaited break where I just play with the language instead so frenziedly work on it.

In term of debugging this is probably it. I really did quite a lot haven't I?