`hacker_rank_2` compiles, but I am starting to understand why it takes so long. There is an insane number of methods for this, more than can be fit in console's buffer. Something is going badly wrong here and I have no idea what.
Actually, as it turns out, it is a pretty good idea to do these HackerRank exercises after all. That there are no more bugs in the compiler is just wistful thinking it seems.

I spent the last hour or so tracking down a bug only to discover that `System.Void` does not equal `unit`.

Let me call it a day here. It is past 6pm and I had enough. I'd rather just shoot demons and watch Symphogear. I'll figure out what is going on here tomorrow.