Based on testing, now I am 100% sure that the slow compilation is because the runtime is dominated by method key lookups.
The bug I just fixed in met statements was related to it. When the environment for it was filtered incorrectly, the runtime absolutely exploded. It took it like 3 extra seconds to produce 40 lines of code.

Now that I know what the problem is, I need to figure out how to deal with it.

Honestly, I have no idea how.

I feel quite dry. It is one thing to say that I need to insert unique IDs, but that will play havoc with rewriting.

If possible I would like to do it without disturbing the existing design too much.

...

I need to look into how hashing work for discriminated unions. I am leaning towards adding a special Hash node or something like that in key places.