I am getting pissed here.
It occured to me that the best thing to use would be an array, but there isn't an array with the same interface. Shit.

I am trying to rush this, but most likely, I will just leave this for tomorrow.

I am going to have to give it the proper respect and work a little on the KeyedCollection class in order to get to know it a little.