With this current redesign the compiler gets a bit over 2x faster, but this is still not enough. The scaling is still superlinear.
The thing that comes to mind is that in nested functions, it also needs to compare the arguments of the arguments of the arguments...

I am going to have to try nodifying the types as well....no, in that case how should I deal with the standard typed expressions. Agh...

I did not think this would be so difficult, but I am going to have to deal with it.

But before I do anything else, let me do instrumentation again.

I won't be able to do much work for the rest today as in an hour I have to leave for other work.