I am going to try redesigning the language completely to improve hashing.
I am gearing up my mind to go through with this.

One thing I want to try before I start with the redesign would be to try making the NodeList. I want to see if it improves the list hashing behavior.

Damn, this is going to be such a pain in the ass. It is making me stress out, but I am going to have to do it. I have no choice in the matter if I want to get the language to a usable state.

But it will take me a while to build up motivation and a while to finish it. I'll give it a few days at the latest.