What a waste of a day. I woke up, watched the 40m Hinton interview by Ng and then I moved stuff around for another 40m.
Then dad brought me a laptop and I spent 4 hours making it work.

And now rather than program, I am agonizing over a design decision. It is nothing like a gear coming loose in my yesterday. What is troubling me today is now that I've decided to make those memo nodes, is how to deal with passing state around.

Consider that standard way I've been doing it up to now, via arguments.

I have like 23 dictionaries just for memoization and probably I am going to add more later.

I have the smart constructors for all the `Ty`s and `Expr`s. That brings the total to 46, or just at 23 if we consider passing only the functions around.

But then the total just goes up again if I consider all the auxiliaries that now aren't pure anymore depite the parser having to access them.

If I was writing the compielr in Spiral this would be no problem - I'd just pack them into a module and then pass that module into the parser and `open` it. The same goes for the codegen.

I've been considering the implications of first class modules for a while now and I am starting to think that the reader pattern of passing the state via an argument is actually an anti-pattern.

It would be better to instantiate the parser combinators via an argument and then export the bound parsers in a module. Then one can simply unpack the module and use them as if they were defined in the same namespace.

There is no doubt that Spiral has some remarkable abilities for syntactic abstraction that F# doesn't.

So what the hell do I do now? It feel like I am being forced to choose between a bunch of bad choices.

But it is good that I picked F#. In Haskell this sort of memoization would not be nearly as effective because of the lack of mutable dictionaries.

I am quite fortunate that I am able to pick this choice at all.

Ah, who needs trully first class modules. This is the correct choice in a ML like language. I should be grateful just that I have a language with lexical scope.

I am not being sarcastic by the way - this is just the reality of programming in mid-2017.

I need a bit ot resolve myself to do it like this. I can already see myself having the entire language in a single 2.5k+ LOC file. Whan it is complete will it grow to 10k or something?

Well, let me give it a shot. I am trying to follow best practices, but the reality won't let me. Maybe I'll surprisingly like it?