I am back in the groove after being distracted by various things for the past week, at no fault of mine.
I finally finished redesigning the language to allow greater memoization of expressions.

The truth is, this is the way it should be organized. Lexical scope demands it and it does not care about putting things in different files, so neither should I. Modules would be an anti pattern here.

I am going to do testing next, and probably I am going to have to nodify TypedExpression as well. I do not this will work that well, but I am hopeful. Maybe I'll try making Node a struct as well to reduce pressure on the GC.

At any rate, let me go through the paces so I can see if I haven't broken everything and then I'll go back to performance testing.