I guess it should not be that surprising that typec_create does not work. First it crashed due to accessing itself, but now I am not sure what to do.
I really having great difficulty trying to represent the recursive relationship using laziness.

...

I think I know. Let me go to the system I used previously, except I will touch it up a bit.

Edit: I forgot to push again it seems.

I really did go all the way back to the previous system. I can't deal with laziness. If I did manage to get it right...

...Actually, I know, I think I got it. Let me try it again.

...

Please spare me, adding lazy nodes just makes the damn thing more complicated. This is a complete waste of time.

The thing is, I still need global dictionaries to connect the damn thing, otherwise there is just no way to do it. And adding laziness to the mix just makes things more complicated in various ways.

Now with that out of the way, let me talk about the last thing that I wanted before I call  it a day.

One of the things that gave me trouble in the codegen is printing of type definitions and I finally know how to do it properly.

I was really enthused by having them all inside those global dictionaries, but that was not such a good idea for the simple reason that it prints out every single type that the partial evaluator ever encountered, and it encountered a lot more than it instantiated.

This causes significant code bloat and some variables use spaces inside of them internally which makes them unprintable.

In the previous versions, the problem that I had with printing types is that sometime iterating over the set and printing the types actually caused new types to be added into the set which caused an exception to be thrown.

I finally know how to do this properly - what I need to do is use a Stack and a HashSet.

I will simply use the HashSet to keep track of all the encountered types and then add then process them using the Stack. That will ensure that the Set does not throw exceptions during processing and I will me able to guarantee completeness.

I am 100% sure that the solution I did a week ago is buggy and poorly made.

It is just that the correct thing to do did not occur to me then, but now that I can take care of it.

Also, at first I was hesitant about packing everything in a single function, but I am actually quite digging this design. I like it. It feels very liberating to have all in one place.

No doubt about it, this is the correct way to program.

Well, I think it is about time to shot some demons. I did well even though in the end I had to rehash old ground. The laziness experiment was a failure and a waste of time, but I needed to try it.

Programming can be like that sometimes, not all of my ideas pan out.

I am out for the day. It is time I shot some demons.

The Nightmare difficulty in Doom is actually not that hard, those fucking secrets though make me waste hours on tracking them down. I think I will just play the rest normally.