Fuck. Git ate my nodified EnvTerm. I have no idea how to revert the damn commits.
But the reason why I reverted is because nodifying EnvTerm actually slowed down the pevaller.

That is not good, it means I won't be able to speed it up by going in this direction. Right now what I need to figure out is how to allow more stack memory so I can get it to compile for n>15.

But let me practice reverting the commits first so this does not happen again. I had no idea the reset would modify even the branches.