Removing adding stuff to the CSE does nothing.
So based on sampling (instrumentation never finishes), what I can tell is that for the the main overheds are F#'s inefficient inbuilt functions and hashing and comparisons on TypedExprs. I did a great job eliminating overhead for Expr and Tys and now the pevaller is much faster.

Maybe I will try nodifying the whole of TypedExprs data structures tomorrow.

And for test33, I want to figure out why it in particular is experiencing superlinear scaling. Is it due to increased cache misses or is it something else? I need to find that out.

Maybe I will profile for memory allocation next.

Actually, let me do that and then I will call it a day. I am getting tired.