As I remove and add layers, the timings remain constant.
There is something I want to do. In test33 I want to print out the number of variables in the environment as the program runs. I absolutely do not understand why I see superlinear scaling in that particular example.

In test34 sure, there could be repeat work, but not in test33.