I've really trimmed it down significantly. I've unrolled a loop that just adds 1 20k times and I see that I get superlinear scaling on it.
Now I know where the problem lies. But what to do about it?

I feel like this is related to the stack usage, but I have no idea how exactly.

Maybe I'll make a tiny intepreter just so I can make sure of this behavior.

I want to see if I get the same super linear scaling with it.