I am finally done with the rewrite. Phew.
Damn, changing a bunch of definitions sure took me a while. I think this change will improve the quality of the partial evaluator.

The reason the TyVV and TyFun had types directly in them was for a pragmatic reason, but I definitely did not count on hashing being exponential because of that.

Thankfully, the latest changes did simplify the compiler a decent bit so this rewrite was definitely worth it.

The next thing then would be to do the testing.