I finally figured out why pattern matching was not working. It turns out tuple indexing was not getting rewritten because I've pared down the destructuring so much. I had to add a call to destructure for that particular work in order to get pattern matching to work again.
Despite how simple the bugs turn out to be in the end, it always takes me hours to work out my perplexity.

Usually I do not program this late, but this bug bothered me so much that I could not rest., so I put in some extra time.

Well, since it is 9:40pm, I think I really will call it a day here.