I measured GC times for the stack based interpreter in that SO post and it does turn out that the GC is reponsible for all the slowness.
That is not a problem is the main partial evaluator, though it is true that it is spending nearly a quarter of the time in GC.

...Ahh, I see it now.

It is not the removing types from TyVV and TyFun is what caused it to go slower.

Rather, by removing those types, it means that it now has to compare them by content in its entirety while previously it could look into the type and shortcut the search.

Though I can't be sure since Ty should be compared last since they are in the last position.

...

One thing I will also need to nodify are strings.

...

No, the above can't possibly be the reason for the slowdown, but I have no idea either way. But the profiler is clear on one this - speed up equality checking.

I will obey.

TypedExprs make me sick to memoize them, but I will do it starting tomorrow. Right now I've already been at it for a full day.

But basically this is the last thing.

After I memoize everything, it will be time to go and do more garden variety optimizations, but at some point if I want performance I need to put my foot down and rewrite Spiral in Spiral.