Time for parse: 00:00:00.1071417 Time for prepass: 00:00:00.0276974 Time for peval was: 00:00:03.7339944 Time for optimization pass was: 00:00:04.1734841 Copied the code to: C:\Users\Marko\Documents\Visual Studio 2015\Projects\ConsoleApplication4\ConsoleApplication4\Program.fs Time for codegen was: 00:00:00.0595930
I made a huge mistake. I had absolutely no idea that the optimization pass took this long.

I am going to have to go to the previous branch and test it there as well. I would not have realized this without PerfView – even when I saw it near the top my instinct was to disbelieve it.