Time for parse: 00:00:00.1095685 Time for prepass: 00:00:00.0276942 Time for peval was: 00:00:02.9145448 Time for optimization pass was: 00:00:00.0386721 Copied the code to: C:\Users\Marko\Documents\Visual Studio 2015\Projects\ConsoleApplication4\ConsoleApplication4\Program.fs Time for codegen was: 00:00:00.0616050
Now this is confusing. It is true that peval takes less here, but it seems it is the optimization pass that takes only a tiny fraction of what it did before.

And the reason is no doubt the memodict.
match memoized_methods.[link_memo.[tag]] with

In here the key is a Node.