I just got up and I already want a break. Yeah, dirty flags would be a great idea if not for one thing – filtration. When I had this huge surge of inpiration yesterday, somehow it completely slipped my mind. I spent the whole night trying to find a way to make it work and in the end while I did manage to conclude that the order of insertion into the environment is something that can be optimized and ordering the keys by their name is not even close to optimal, in order to avoid renaming nested structures it would literally have to be perfect.
This happens sometimes. I spent the whole night tossing and turning the new insight around only to have it fall apart. Programming is risky. I won't be able to squish the exponential scaling in renmaing for the parser examples this way.

What a waste. Let me finish the Lisp book.