"1:55pm. No, I just don't see it. And not lifting procedure up to the top level has its own issues, but the benefits is not something I will see while doing combinators.
...It can't be done.

Renaming isn't something I put in a trivial amount of thought into either. Do you remember the March, April and May months and how long it took me to figure out the basic scheme of renaming I am using currently?

It literally took me months and it one of the great insights of the Spiral language on how to memoize join points.

I'll convert strings to ints, and make the rename pass a bit faster, but that is it.

2:05pm. Actually last night before I took a bath I even decided to give up on renaming before I got a surge of inspiration all of a sudden. So let me go to that...

...No. That is the wrong choice.

I won't give up on this feeling. It is not how I intended to do things, but I will let it blaze for as long as it wants until it either crystalizes or burns out.

I am not in a hurry here. If it turns out to be nothing then I will have proven to myself that it is impossible to root out the hotspot and will be able to move forward with a clear conscience."

I am finally done with the Lisp book and it was a complete waste of time. I am not having anything to do with Lisp again. It is not simple, nor is it elegant, nor is it fast.

Maybe I should try looking for something similar except in ML.

http://www.paulgraham.com/thist.html

I am also reading on the history of Lisp and find it hillarious that even during early 80s there was a debate between dynamic and lexical scoping proponents in the Lisp circles. Lispers arguing that lexical scoping could not be done because dynamic scoping would be too slow honestly boggles my mind.

...

I am still gripped by obsession over the renaming issue.

I want to see how MLton does things like specialization. 5-10m for 140k LOC is not too horrible. It comes to about 350 LOCs per second.

I think this is actually about par for Spiral, the tuple parser example I am trying to resolve is a special case that static languages do not even have.

MLton's fatal flaw is that it does whole program optimization. This is something that is not a problem to Spiral and the road to success for the language most likely lies in taking advantage of that.

Spiral is unrivaled, and it could only have been written in a ML variant like F#.

Though it is painful, I am think that since I am in the studying groove I should check out the Tiger book as well. That might be more up my alley.

At the very least, I'll browse MLton's source and maybe look for partial evaluators for ML.