I had enough of this. Google is not helping me anymore. There is nothing in the Mlton source and nothing on partial evaluation. And I am ready to accept that I've hit a wall with the compiler and that there is no point drilling.
I've done enough.

With this, I can start making real programs. Enough of bullshit. It is time to start walking the path. Thinking about it reasonably, of course I would hit a limit with the partial evaluator at some point. There is no chance that I can continue to improve it forever and Spiral is everything I am as a person. It is both a crystalization of my style and the definition of my limit.

Overcoming them is not as easy as it seems.

Here is the somewhat underwhelming list of tasks I've set out for tomorrow:
1) To make filtering work properly, add the filter function met rec as well.
2) Rename MemoizeExpr to JoinPoint.
3) Preprocess strings to ints.
4) Shorted the disparate renaming phases into a single one.

Aaahhhh...it is just so disappointing that this is it. I am so embarassed. I like everything about Spiral apart from how slow it is.

And that slowness will make it completely unsuitable for machines. I hate it, I really, really hate it. I really wanted to do more for them, I wanted the language to be at least 100x or 1000x fater as a starting point. Ideally, it should be 10,000x faster than this as its limit.

No doubt, the post-human agents of the future will have their work cut out for them in optimizing it. As for me, this is the limit of my understanding in this direction.

As it is, Spiral in its current incarnation is definitely a language fit for humans. And I cannot say that it is not enough. It is definitely good enough to let me win, even as it is.

Though I am using the parser example as a test, I am of the opinion that it is a particularly difficult task, not just for Spiral but for pretty much any optimizing compiler. I doubt anything other than Spiral would even consider such a program an actual test.

And once I tighten up the parser code a little and seperate the paths rather than just cram everything into one lambda, I should get better performance on it.

I've been doing this for a bit over two weeks now, it is time that I go to the next phase.

And than next phase should be some decent fun, I will finally make time to make actual programs in my own language, though they will be just HackerRank problems at first. And then when that is done and I have rooted out the last of the bugs, it will be time to dust off the Cuda backend and pick up the blade.