""7:10pm. http://dblp.uni-trier.de/pers/hd/a/Asai:Kenichi
Finally found where this guy keeps his papers.
For some reason the links on his personal page are broken.

7:20pm. Change of plans...http://library.readscheme.org/page11.html

The scheme library's paper titles on reflection are suggestive. I was right that there is a link between reflection and partial evaluation. I need to read some of Kenichi Asai's work. Rather than start work tomorrow, I think I will lengten my studies instead to read the papers
That reflective languages are uncompilable and yet, partially evaluable...

This brings an idea to my mind. A single question: What if I partially evaluated the partial evaluator with respect to the program being compiled?

Those nodification optimizations, I said they were a beginning of a pool allocator, but I wonder if that is really true? I obsessed and ran in circles today over renaming, but if I step back and think, I need to ask myself, do I really have the big picture in my mind or did I convince myself that what Spiral can do here is all that partial evaluation can do.

I assumed that pevalling the pevaller would make it make it faster (though I would need to rewrite it,) and obviously, pevalling a program would compile it, but I also assumed pevalling a pevaller with the input given to it would do nothing other than simply pevalling a program, that is if the input was static.

But if I think about it more reasonably, that can't be true can it?

No, it can't. What it would do instead is optimize the pevaller for compilation speed by specializing it for a particular program. I had things mixed up in my mind."

I wanted to post the above exerpt from my journal yesterday, but I was too tired. I actually went to bed 2.5 hours earlier than usual as I was so tired yesterday.

I had another big insight into and I can see the big picture now. It is as I said, to optimize compilation times I need to specialize the partial evaluator to code fragments. To draw out the full power of that technique, I would need true first class staging, meaning the partial evaluator should be able to run its own code and call itself.

Right now I can only evaluate a single level level of staging. In other words, the partial evaluator needs to be reflective.

I do not really understand all the steps, and how things would turn out with reflective staging and all the steps needed to make the technique work, but the answer cannot be anything other than this.

Yesterday was just so difficult - I came all this way and did all this work, and yet it was enough.

I implemented the partial evaluator and when I stood back and admired my work I found it perfect.

Yet, it was perfection that needed to be surpassed.

Looking back on my life, I've never ever managed to defeat these kinds of scenario, but it seems that for the first time, there is hope - I will be able to move forward and surmount even perfection.

Putting it into context, for supporting the broadest range of possible partial evaluation scenarios, the current evalutor really can't be improved much in terms of performance. Not in F# and .NET - fundamentally, its design is sound.

My rewriting in itself, I might be able to trim some fat, but even if I compiled to a more bare language and did my own memoizing memory allocation that would not be without its drawbacks.

So forgetting the dumb idea of doing memoizing allocation which would be more trouble than it is worth, by optimizing the evaluator properly and getting rid of all the dumb intermediates I have for allocations, I do think I could speed it up by 10x.

Though that entails having to redo not just the pevaller, but also the data structures it uses, so I am not gung ho about doing this work now.

Even worse, if I really resolve myself to go all the way, I am going to have to make partially static variants of not just tuples, but also immutable maps and dictionaries.

The vision of this how the futural Spiral will look like is that it will have layered staging, and compilation times in libraries will be put down by specilizing the compiler to them.

This vision of the future is something that is currently far afield - the only way to get generic libraries now is in static languages with generics and dynamic languages.

In all cases, the libraries either have to be recompiled from the top. In no cases is the compiler ever fused to the library directly - something like that would be positively demonic.

It is without a doubt the fragment of the true future, but it is not even on the drawing board.

Let me close this long rable by saying that I am not going to do it right now nor for the foreseable future.

I certainly do not need those features right now, and neither do I need the Spiral partial evaluator to be faster. It is exactly how I said a few days ago - the state of the system is certainly well and sufficient to allow me to succeed at what I intend to do.

And the reality of what drove me to pursue this more than two and a half years ago has not changed - I need power and I need help. If my agents come to require a much better language than what I have on offer, then they are going to have to give me a hand at this.

There is only so much one man can do on his own.

The basic single layer online partial evaluator that I have now is enough.

What I will do is write more code in it and crystalize my style.

For today, I think I will tie up the loose ends and read some of the papers by Kenichi Asai on reflection and partial evaluation. The dbpl library really should be my goto for computer science material.